smartchat-agent/pyproject.toml
[project]
name = "smartchat-agent"
version = "0.1.0"
description = "LangGraph AI agent for SmartChat - multi-person messaging and conversation intelligence"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.4.1",
    "langchain>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-core>=0.3.0",
    "supabase>=2.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "apscheduler>=3.10.0,<4.0.0",
    "python-dateutil>=2.8.0",
    "sqlalchemy>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"
